admin_tools_stats/requirements.txt
python-dateutil>=1.5,<2.0
django-jsonfield==0.8
django-qsstats-magic>=0.6.1
django-chart-tools>=0.2.1
python-memcached>=1.47
django-cache-utils
django-admin-tools==0.4.1
git+git://github.com/Star2Billing/switch2bill-common.git